backend/requirements.txt
aiohttp==3.9.5
aiomysql==0.2.0
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.1.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
fastapi==0.110.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
makefun==1.15.2
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
pandas==2.2.2
pwdlib==0.2.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
requests==2.32.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlmodel==0.0.16
starlette==0.37.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.29.0
WTForms==3.1.2
yarl==1.9.4
tenacity==8.3.0
locust==2.27.0
networkx==3.3
matplotlib==3.9.0
cachetools==5.3.3
redis==5.0.7
passlib==1.7.4
pytest==8.2.0
httpx==0.27.0
trio==0.26.0
uvloop==0.20.0
httptools==0.6.1
pytest-asyncio==0.23.7
testcontainers==4.8.1
factory_boy==3.3.1
Faker==29.0.0
